ajout du -t

+++ srcs/parsing/ft_flush_time_flag_callback.c
@@ -6,7 +6,7 @@
 /*   By: sclolus <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/23 20:09:15 by sclolus           #+#    #+#             */
-/*   Updated: 2017/11/24 00:52:51 by sclolus          ###   ########.fr       */
+/*   Updated: 2017/12/01 22:31:27 by sclolus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,6 @@
 int32_t	ft_flush_time_flag_callback(t_script_info* script_info)
 {
 	script_info->flags.bits.flush_time = 1;
-	script_info->flush_time = ft_atol(g_optarg);
+	script_info->flush_time = (int64_t)ft_atol(g_optarg);
 	return (0);
 }
